meme wall features and other updates
